feat(settings/native): moved everything to settings feature and added bottom inset
