+++ react/features/settings/components/native/FormRow.js
@@ -3,7 +3,7 @@
 import React, { Component } from 'react';
 import { Text, View } from 'react-native';
 
-import { translate } from '../../../../../base/i18n';
+import { translate } from '../../../base/i18n';
 
 import styles, { ANDROID_UNDERLINE_COLOR, PLACEHOLDER_COLOR } from './styles';
