--- react/features/welcome/components/native/settings/components/index.js
@@ -1,1 +1,0 @@
-export { default as SettingsView } from './SettingsView';
